Non OOP, no UI version
This version has functions, but they do not belong to any class. The game runs for 100 turns flat and does not have a check mechanism to see if the game is over. It also cannot understand when the game is over. İt is controlled by entering 8 for up, 6 for right, 2 for down, and 4 for left as these numbers in the calculator are shaped in a four-sided compass-like shape.