With OOP, no UI
The code now checks if the game is over and whether the player has won. This version still has the 8,6,2,4 arrangement to command the player's movements. It completely works and asks for the right input if the entered input is not fit to be manipulated. How information (the game's table) is presented is optimized for the calculator's screen.