OOP with PyGame UI
There is a UI to display the game's table in a new window. The game is now played with the arrow keys on the keyboard as it would in a professional game setting. It can track the user's progress and see if they have won or lost. In case of a loss, the system displays it to the user and automatically shuts down. As the number of digits in the numbers increases, their size decreases gradually to let them fit on the screen. It is better optimized for use in a computer.